requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-crispy-forms==1.8.1
djangorestframework==3.11.0
idna==2.8
Pillow==6.2.1
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
